Split KNRM into different files
